requirements.txt
Django==4.2.0
djangorestframework==3.14.0
openai==1.3.0
pillow==10.0.0
channels==4.0.0
celery==5.3.0 
numpy==1.26.4